+++ Tests/AfterTraining/test_after_training.py
@@ -33,7 +33,8 @@ from InnerEye.Common.fixed_paths_for_tests import full_ml_test_data_path
 from InnerEye.Common.output_directories import OutputFolderForTests
 from InnerEye.Common.spawn_subprocess import spawn_and_monitor_subprocess
 from InnerEye.ML.common import DATASET_CSV_FILE_NAME, ModelExecutionMode
-from InnerEye.ML.deep_learning_config import CHECKPOINT_FOLDER
+from InnerEye.ML.deep_learning_config import CHECKPOINT_FOLDER, ModelCategory
+from InnerEye.ML.reports.notebook_report import get_html_report_name
 from InnerEye.ML.utils.image_util import get_unit_image_header
 from InnerEye.ML.utils.io_util import zip_random_dicom_series
 from InnerEye.Scripts import submit_for_inference
@@ -197,6 +198,9 @@ def test_expected_cv_files_classification(test_output_dirs: OutputFolderForTests
     # We should not have any ensemble metrics in CV folder
     for mode in [ModelExecutionMode.TEST, ModelExecutionMode.TRAIN, ModelExecutionMode.VAL]:
         assert not _check_presence_cross_val_metrics_file(ENSEMBLE_SPLIT_NAME, mode, available_files)
+    crossval_report_name = f"{ModelCategory.Classification.value}_crossval"
+    crossval_report_file = f"{CROSSVAL_RESULTS_FOLDER}/{get_html_report_name(crossval_report_name)}"
+    assert crossval_report_file in available_files
 
 
 @pytest.mark.skipif(common_util.is_windows(), reason="Too slow on Windows")
@@ -275,8 +279,7 @@ def test_register_and_score_model(test_output_dirs: OutputFolderForTests) -> Non
     assert_nifti_content(str(expected_segmentation_path), expected_shape, image_header, [3], np.ubyte)
 
 
-# @pytest.mark.after_training_2node
-@pytest.mark.skip("2 nodes training hangs with PL 1.2.7")
+@pytest.mark.after_training_2node
 def test_training_2nodes(test_output_dirs: OutputFolderForTests) -> None:
     """
     Test if a job running on 2 nodes trains correctly.